+++ phenodata/dwd/pheno.py
@@ -84,13 +84,11 @@ class DwdPhenoData(object):
         Apply a bunch of filters to the result data.
         """
 
-        observations = self.query(partition=options['partition'], files=options['files'])
+        # Acquire data
+        observations = self.query(partition=options['partition'], files=options['filename'])
 
-        observations = self.flux(
-            observations,
-            stations=options['stations'], regions=options['regions'],
-            species=options['species'], phases=options['phases'],
-            years=options['years'], forecast=options['forecast'])
+        # Filter data
+        observations = self.flux(observations, criteria=options, forecast=options['forecast'])
 
         return observations
 
@@ -107,7 +105,7 @@ class DwdPhenoData(object):
         logger.info('Starting data acquisition')
 
         # Search FTP server
-        paths = self.scan_files(partition, files=files, field='url')
+        paths = self.scan_files(partition, include=files, field='url')
 
         # The main DataFrame object
         results = pd.DataFrame()
@@ -115,16 +113,16 @@ class DwdPhenoData(object):
         # Load multiple files into single DataFrame
         for path in paths:
 
+            # Skip invalid files
+            if 'Kulturpflanze_Ruebe_akt' in path:
+                logger.warning('Skipping file "{}" due to invalid header format (all caps)'.format(path))
+                continue
+
             # Acquire DataFrame from CSV data
             data = self.cdc.get_dataframe(path, coerce_int=True)
 
             # Sanity checks
             if data is None:
-                continue
-
-            # Skip invalid files
-            if 'Kulturpflanze_Ruebe_akt' in path:
-                logger.warning('Skipping file "{}" due to invalid header format (all caps)'.format(path))
                 continue
 
             # Coerce "Eintrittsdatum" column into date format
@@ -143,12 +141,18 @@ class DwdPhenoData(object):
 
         return results
 
-    def flux(self, results, stations=None, regions=None, species=None, phases=None, years=None, forecast=False):
+    def flux(self, results, criteria=None, forecast=False):
         """
         The flux compensator. All filtering on the DataFrame takes places here.
         """
 
         logger.info('Flux compensator: Filter and transform data')
+
+        criteria = criteria or {}
+
+        # Sanity checks
+        if results is None:
+            return
 
         # Filter DataFrame
         # https://pythonspot.com/pandas-filter/
@@ -156,11 +160,27 @@ class DwdPhenoData(object):
 
         # Build "boolean indexing" filter expression from multiple criteria
         # https://pandas.pydata.org/pandas-docs/stable/indexing.html#boolean-indexing
+        isin_map = {
+            'year': results.Referenzjahr,
+            'quality-level': results.Qualitaetsniveau,
+            'quality-byte': results.Eintrittsdatum_QB,
+            'station-id': results.Stations_id,
+            'species-id': results.Objekt_id,
+            'phase-id': results.Phase_id,
+        }
+
+        # Lowlevel filtering based on IDs
+        # For each designated field, add ``.isin`` criteria to "boolean index" expression
         expression = True
-        if stations:
-            expression &= results.Stations_id.isin(stations)
-        if years:
-            expression &= results.Referenzjahr.isin(years)
+        for key, reference in isin_map.items():
+            if criteria[key]:
+                expression &= reference.isin(criteria[key])
+
+
+        # Humanized filtering based on merged/joined DataFrames
+        # TODO: For each designated field, add ``.str.contains('|'.join(patterns)`` criteria to "boolean index" expression
+        # https://stackoverflow.com/questions/12065885/filter-dataframe-rows-if-value-in-column-is-in-a-set-list-of-values/26724725#26724725
+
 
         # Apply filter expression to DataFrame
         if type(expression) is not bool:
@@ -168,7 +188,7 @@ class DwdPhenoData(object):
 
         return results
 
-    def scan_files(self, partition, files=None, field=None):
+    def scan_files(self, partition, include=None, field=None):
         """
         Scan upstream files in three-level directory hierarchy.
         """
@@ -179,7 +199,7 @@ class DwdPhenoData(object):
         # Query FTP server for files
         entries = self.cdc.ftp.scan_files(
             url, subdir=partition,
-            include=files,
+            include=include,
             include_base=['PH_(Sofort|Jahres)melder.+\.txt'],
             exclude_base=['PH_Beschreibung', 'Spezifizierung'],
         )